// src/types/dx-preview.ts
export enum ValidationStatus {
  SUCCESS = 'SUCCESS',
  ERROR_NOT_FOUND = 'ERROR_NOT_FOUND',
  ERROR_FORBIDDEN = 'ERROR_FORBIDDEN',
  ERROR_BAD_REQUEST = 'ERROR_BAD_REQUEST',
  ERROR_SERVER = 'ERROR_SERVER',
  ERROR_FORMAT_UNSUPPORTED = 'ERROR_FORMAT_UNSUPPORTED'
}